Move tests to Wrangler unit tests
